services/scraper/requirements.txt
scrapy==2.11.0
beautifulsoup4==4.12.2
requests==2.31.0
psycopg2-binary==2.9.9
redis==5.0.1
pika==1.3.2
python-dotenv==1.0.0
APScheduler==3.10.4
structlog==23.2.0
httpx==0.25.2
